move the queue bar to appframe folder
