+++ src/components/safe-apps/AppFrame/index.tsx
@@ -8,16 +8,16 @@ import useSafeInfo from '@/hooks/useSafeInfo'
 import { useSafeAppFromBackend } from '@/hooks/safe-apps/useSafeAppFromBackend'
 import useChainId from '@/hooks/useChainId'
 import { isSameUrl } from '@/utils/url'
+import { isMultisigDetailedExecutionInfo } from '@/utils/transaction-guards'
 import useThirdPartyCookies from './useThirdPartyCookies'
 import useAppIsLoading from './useAppIsLoading'
 import useAppCommunicator, { CommunicatorMessages } from './useAppCommunicator'
-import useTxModal from '../SafeAppsTxModal/useTxModal'
 import { ThirdPartyCookiesWarning } from './ThirdPartyCookiesWarning'
 import SafeAppsTxModal from '../SafeAppsTxModal'
+import useTxModal from '../SafeAppsTxModal/useTxModal'
 import SafeAppsSignMessageModal from '../SafeAppsSignMessageModal'
 import useSignMessageModal from '../SignMessageModal/useSignMessageModal'
-import { isMultisigDetailedExecutionInfo } from '@/utils/transaction-guards'
-import TransactionQueueBar from '@/components/safe-apps/TransactionQueueBar'
+import TransactionQueueBar from './TransactionQueueBar'
 
 import css from './styles.module.css'
